chore(infra): move to turbopack (without jquery vendor source => simulator doesn't work yet)
