+++ package.json
@@ -4,10 +4,9 @@
   "private": true,
   "type": "module",
   "scripts": {
-    "dev": "next dev",
-    "preview": "next build && next start",
-    "build": "next build",
-    "start": "next start",
+    "dev": "next dev --turbo",
+    "build": "next build --turbo",
+    "preview": "next build --turbo && cp -r public .next/standalone/ && cp -r .next/static .next/standalone/.next/ && bun .next/standalone/server.js",
     "db:migrate": "drizzle-kit migrate",
     "db:migrate-no-legacy": "./scripts/migrate-no-legacy.sh",
     "db:generate": "drizzle-kit generate",
