+++ src/app/landing/_shared/icons.tsx
@@ -1,10 +1,10 @@
-import CubeIcon from '@/../public/landing/cube.svg?inline'
-import KeyboardIcon from '@/../public/landing/keyboard.svg?inline'
-import LeaderboardIcon from '@/../public/landing/leaderboard.svg?inline'
-import ResultIcon from '@/../public/landing/result.svg?inline'
-import ScrambleIcon from '@/../public/landing/scramble.svg?inline'
-import ShareIcon from '@/../public/landing/share.svg?inline'
-import StopwatchIcon from '@/../public/landing/stopwatch.svg?inline'
+import CubeIcon from '@/../public/landing/icons/cube.icon.svg'
+import KeyboardIcon from '@/../public/landing/icons/keyboard.icon.svg'
+import LeaderboardIcon from '@/../public/landing/icons/leaderboard.icon.svg'
+import ResultIcon from '@/../public/landing/icons/result.icon.svg'
+import ScrambleIcon from '@/../public/landing/icons/scramble.icon.svg'
+import ShareIcon from '@/../public/landing/icons/share.icon.svg'
+import StopwatchIcon from '@/../public/landing/icons/stopwatch.icon.svg'
 
 export {
   CubeIcon,
