+++ src/frontend/ui/icons.tsx
@@ -1,41 +1,41 @@
-import Discipline3by3Icon from '@/../public/icons/discipline-3by3.svg?inline'
-import Discipline2by2Icon from '@/../public/icons/discipline-2by2.svg?inline'
-import AllContestsIcon from '@/../public/icons/all-contests.svg?inline'
-import ArrowBackUpIcon from '@/../public/icons/arrow-back-up.svg?inline'
-import ArrowRightIcon from '@/../public/icons/arrow-right.svg?inline'
-import AvatarIcon from '@/../public/icons/avatar.svg?inline'
-import CheckIcon from '@/../public/icons/check.svg?inline'
-import CloseIcon from '@/../public/icons/close.svg?inline'
-import DashboardIcon from '@/../public/icons/dashboard.svg?inline'
-import DiscordIcon from '@/../public/icons/discord.svg?inline'
-import ExclamationCircleIcon from '@/../public/icons/exclamation-circle.svg?inline'
-import GoogleIcon from '@/../public/icons/google.svg?inline'
-import LeaderboardIcon from '@/../public/icons/leaderboard.svg?inline'
-import ChevronLeftIcon from '@/../public/icons/chevron-left.svg?inline'
-import ChevronRightIcon from '@/../public/icons/chevron-right.svg?inline'
-import ChevronUpIcon from '@/../public/icons/chevron-up.svg?inline'
-import ChevronDownIcon from '@/../public/icons/chevron-down.svg?inline'
-import LinkedinIcon from '@/../public/icons/linkedin.svg?inline'
-import LogoutIcon from '@/../public/icons/logout.svg?inline'
-import MenuIcon from '@/../public/icons/menu.svg?inline'
-import OngoingContestIcon from '@/../public/icons/ongoing-contest.svg?inline'
-import PlaySkipBackIcon from '@/../public/icons/play-skip-back.svg?inline'
-import PlaySkipForwardIcon from '@/../public/icons/play-skip-forward.svg?inline'
-import PlaybackIcon from '@/../public/icons/play-back.svg?inline'
-import PlayForwardIcon from '@/../public/icons/play-forward.svg?inline'
-import PlayIcon from '@/../public/icons/play.svg?inline'
-import ShareIcon from '@/../public/icons/share.svg?inline'
-import SortIcon from '@/../public/icons/sort.svg?inline'
-import StopIcon from '@/../public/icons/stop.svg?inline'
-import GithubIcon from '@/../public/icons/github.svg?inline'
-import EllipsisIcon from '@/../public/icons/ellipsis.svg?inline'
-import PlusIcon from '@/../public/icons/plus.svg?inline'
-import MinusIcon from '@/../public/icons/minus.svg?inline'
-import SettingIcon from '@/../public/icons/setting.svg?inline'
-import CodeIcon from '@/../public/icons/code.svg?inline'
-import AverageIcon from '@/../public/icons/average.svg?inline'
-import SingleIcon from '@/../public/icons/single.svg?inline'
-import VscubingIcon from '@/../public/icons/vscubing.svg?inline'
+import Discipline3by3Icon from '@/../public/icons/discipline-3by3.icon.svg'
+import Discipline2by2Icon from '@/../public/icons/discipline-2by2.icon.svg'
+import AllContestsIcon from '@/../public/icons/all-contests.icon.svg'
+import ArrowBackUpIcon from '@/../public/icons/arrow-back-up.icon.svg'
+import ArrowRightIcon from '@/../public/icons/arrow-right.icon.svg'
+import AvatarIcon from '@/../public/icons/avatar.icon.svg'
+import CheckIcon from '@/../public/icons/check.icon.svg'
+import CloseIcon from '@/../public/icons/close.icon.svg'
+import DashboardIcon from '@/../public/icons/dashboard.icon.svg'
+import DiscordIcon from '@/../public/icons/discord.icon.svg'
+import ExclamationCircleIcon from '@/../public/icons/exclamation-circle.icon.svg'
+import GoogleIcon from '@/../public/icons/google.icon.svg'
+import LeaderboardIcon from '@/../public/icons/leaderboard.icon.svg'
+import ChevronLeftIcon from '@/../public/icons/chevron-left.icon.svg'
+import ChevronRightIcon from '@/../public/icons/chevron-right.icon.svg'
+import ChevronUpIcon from '@/../public/icons/chevron-up.icon.svg'
+import ChevronDownIcon from '@/../public/icons/chevron-down.icon.svg'
+import LinkedinIcon from '@/../public/icons/linkedin.icon.svg'
+import LogoutIcon from '@/../public/icons/logout.icon.svg'
+import MenuIcon from '@/../public/icons/menu.icon.svg'
+import OngoingContestIcon from '@/../public/icons/ongoing-contest.icon.svg'
+import PlaySkipBackIcon from '@/../public/icons/play-skip-back.icon.svg'
+import PlaySkipForwardIcon from '@/../public/icons/play-skip-forward.icon.svg'
+import PlaybackIcon from '@/../public/icons/play-back.icon.svg'
+import PlayForwardIcon from '@/../public/icons/play-forward.icon.svg'
+import PlayIcon from '@/../public/icons/play.icon.svg'
+import ShareIcon from '@/../public/icons/share.icon.svg'
+import SortIcon from '@/../public/icons/sort.icon.svg'
+import StopIcon from '@/../public/icons/stop.icon.svg'
+import GithubIcon from '@/../public/icons/github.icon.svg'
+import EllipsisIcon from '@/../public/icons/ellipsis.icon.svg'
+import PlusIcon from '@/../public/icons/plus.icon.svg'
+import MinusIcon from '@/../public/icons/minus.icon.svg'
+import SettingIcon from '@/../public/icons/setting.icon.svg'
+import CodeIcon from '@/../public/icons/code.icon.svg'
+import AverageIcon from '@/../public/icons/average.icon.svg'
+import SingleIcon from '@/../public/icons/single.icon.svg'
+import VscubingIcon from '@/../public/icons/vscubing.icon.svg'
 
 import { isDiscipline } from '@/types'
 import { cn } from '@/frontend/utils/cn'
